--- packages/stage-ui/src/components/HoloCard/index.ts
@@ -1,1 +1,0 @@
-export { default as HoloCard } from './HoloCard.vue'
