+++ packages/stage-ui/src/components/Physics/CursorFloating.vue
