+++ packages/stage-ui/src/components/Physics/index.ts
@@ -1,1 +1,2 @@
+export { default as CursorFloating } from './CursorFloating.vue'
 export { default as CursorMomentum } from './CursorMomentum.vue'
