+++ packages/stage-ui/src/components/index.ts
@@ -4,7 +4,6 @@ export { default as Collapsable } from './Collapsable.vue'
 export * from './Form'
 export * from './Gadgets'
 export * from './Graphics'
-export * from './HoloCard'
 export { default as Live2DCanvas } from './Live2D/Canvas.vue'
 export { default as Live2DModel } from './Live2D/Model.vue'
 export * from './Menu'
